Almost done server

+++ lab4/MonitorTotals.java
@@ -1,17 +1,18 @@
+import java.util.*;
 public class MonitorTotals {
 	public String recordType;
-	public ArrayList<Integer> msgTotals;
-	public ArrayList<Integer> userTotals;
-	public ArrayList<Integer> newMsgTotals;
+	public ArrayList<Long> msgTotals;
+	public ArrayList<Long> userTotals;
+	public ArrayList<Long> newMsgTotals;
 
-	public MonitorTotals(String recordType, ArrayList<Integer> msgTotals, ArrayList<Integer> userTotals, ArrayList<Integer> newMsgTotals) {
+	public MonitorTotals(String recordType, ArrayList<Long> msgTotals, ArrayList<Long> userTotals, ArrayList<Long> newMsgTotals) {
 		this.recordType = recordType;
 		this.msgTotals = msgTotals;
 		this.userTotals = userTotals;
 		this.newMsgTotals = newMsgTotals;
 	}
 
-	public void updateMonitorTotals(int msgTotal, int userTotal, int newMsgTotal) {
+	public void updateMonitorTotals(long msgTotal, long userTotal, long newMsgTotal) {
 		msgTotals.add(msgTotal);
 		userTotals.add(userTotal);
 		newMsgTotals.add(newMsgTotal);
